# Move broadcast to iteration method