--- a/REQUIRE
+++ b/REQUIRE
@@ -5,6 +5,6 @@ WeakRefStrings 0.5.7
 IteratorInterfaceExtensions 0.1.0
 TableTraits
 Tables
-StructArrays 0.3.1
+StructArrays 0.3.2
 DataValues 0.4.6
 TableTraitsUtils